Extract "uniswap" command subtree configuration
With the CLI complexity growing, putting every command configuration
directly into `index.ts` would cause the file to grow to an unmanageable
size.
